Merge with master has not been done since months...

+++ utils/run_test.sh
@@ -56,6 +56,18 @@ cat >> "${fname}" << EOF
 EOF
 }
 
+run_test_filter()
+{
+    if [ -n "${RUN_TEST_FILTER}" ]; then
+        if ! [[ "${COMP} $*" =~ ${RUN_TEST_FILTER} ]]; then
+            echo "Skipping test: $*"
+            return 1
+        fi
+    fi
+
+    return 0
+}
+
 run_test()
 {
     local in="$*"
@@ -64,17 +76,15 @@ run_test()
     b="${b//;/-}"
     b="${b//\"/-}"
 
-    if [ -n "${RUN_TEST_FILTER}" ]; then
-        if ! [[ "$*" =~ ${RUN_TEST_FILTER} ]]; then
-            echo "Skipping test: $in"
-            return
-        fi
+    if ! run_test_filter "$*"; then
+        return
     fi
+
     export D_LOG_FILE="/tmp/daos_${b}-${log_num}.log"
-    echo "Running $* with log file: ${D_LOG_FILE}"
 
     export TNAME="${b}-${log_num}"
 
+    echo "Running $* with log file: ${D_LOG_FILE}"
     # We use grep to filter out any potential "SUCCESS! NO TEST FAILURES"
     #    messages as daos_post_build.sh will look for this and mark the tests
     #    as passed, which we don't want as we need to check all of the tests
@@ -132,23 +142,26 @@ if [ -d "/mnt/daos" ]; then
         run_test src/common/tests/btree.sh dyn perf ukey -s 20000
         BTREE_SIZE=20000
 
+        COMP="UTEST_vos_nvme"
+        if run_test_filter "sudo -E ${SL_PREFIX}/bin/vos_tests -a"; then
+            cat /proc/meminfo
+            # Setup AIO device
+            AIO_DEV=$(mktemp /tmp/aio_dev_XXXXX)
+            dd if=/dev/zero of="${AIO_DEV}" bs=1G count=4
+
+            # Setup daos_nvme.conf
+            NVME_CONF="/mnt/daos/daos_nvme.conf"
+            cp -f src/vos/tests/daos_nvme.conf ${NVME_CONF}
+            sed -i "s+\"filename\": \".*\"+\"filename\": \"${AIO_DEV}\"+g" ${NVME_CONF}
+
+            export VOS_BDEV_CLASS="AIO"
+            run_test "sudo -E ${SL_PREFIX}/bin/vos_tests" -a
+
+            rm -f "${AIO_DEV}"
+            rm -f "${NVME_CONF}"
+        fi
+
         COMP="UTEST_vos"
-        cat /proc/meminfo
-        # Setup AIO device
-        AIO_DEV=$(mktemp /tmp/aio_dev_XXXXX)
-        dd if=/dev/zero of="${AIO_DEV}" bs=1G count=4
-
-        # Setup daos_nvme.conf
-        NVME_CONF="/mnt/daos/daos_nvme.conf"
-        cp -f src/vos/tests/daos_nvme.conf ${NVME_CONF}
-        sed -i "s+\"filename\": \".*\"+\"filename\": \"${AIO_DEV}\"+g" ${NVME_CONF}
-
-        export VOS_BDEV_CLASS="AIO"
-        run_test "sudo -E ${SL_PREFIX}/bin/vos_tests" -a
-
-        rm -f "${AIO_DEV}"
-        rm -f "${NVME_CONF}"
-
         run_test src/vos/tests/evt_stress.py
         run_test src/vos/tests/evt_stress.py --algo dist_even
         run_test src/vos/tests/evt_stress.py --algo soff
@@ -162,6 +175,7 @@ if [ -d "/mnt/daos" ]; then
                                           --show-reachable=yes \
                                           --num-callers=20 \
                                           --error-limit=no \
+                                          --fair-sched=try \
                                           --suppressions=${VALGRIND_SUPP} \
                                           --gen-suppressions=all \
                                           --error-exitcode=42 \
@@ -185,7 +199,6 @@ if [ -d "/mnt/daos" ]; then
 
     COMP="UTEST_vos"
     run_test "${SL_PREFIX}/bin/vos_tests" -A 500
-    run_test "${SL_PREFIX}/bin/vos_tests" -n -A 500
     COMP="UTEST_vos"
     cmd="-c pool -w key@0-4 key@3-4 -R key@3-3 -w key@5-4 -R key@5-3 -a -i -d -D"
     run_test "${SL_PREFIX}/bin/vos_tests" -r "\"${cmd}\""
@@ -200,6 +213,13 @@ if [ -d "/mnt/daos" ]; then
     COMP="UTEST_vea"
     run_test "${SL_PREFIX}/bin/vea_ut"
     run_test "${SL_PREFIX}/bin/vea_stress -d 60"
+    # regression test for DAOS-12256
+    COMP="UTEST_vea_debug"
+    export D_LOG_MASK=DEBUG
+    export DD_SUBSYS=all
+    export DD_MASK=all
+    run_test "${SL_PREFIX}/bin/vea_ut"
+    unset D_LOG_MASK DD_SUBSYS DD_MASK
 
     COMP="UTEST_bio"
     run_test "${SL_BUILD_DIR}/src/bio/smd/tests/smd_ut"
@@ -255,10 +275,10 @@ if [ -d "/mnt/daos" ]; then
     COMP="UTEST_csum"
     run_test "${SL_PREFIX}/bin/srv_checksum_tests"
     run_test "${SL_PREFIX}/bin/pool_scrubbing_tests"
+    run_test "${SL_PREFIX}/bin/rpc_tests"
 
     COMP="UTEST_vos"
     run_test src/vos/tests/evt_ctl.sh
-    run_test src/vos/tests/evt_ctl.sh pmem
     unset USE_VALGRIND
     unset VALGRIND_SUPP
